Move to a more modular structure
This is in preparation for a version to come which will offer,
at least, another SVG version.  I am not certain exactly when that will
happen but this will make it easier.

I do not intend to change the extant outputs so this can be relied upon
to produce expected long-term identifiers